no message
git-svn-id: svn://rdo.rk9.bmstu.ru/var/svn/root@1186 23f926e6-80f1-49ca-85a0-60d4a27bb3ac

+++ trunk/rp_method/rdoprocess_object_flowchart.h
@@ -95,7 +95,7 @@ protected:
 	bool            trash_over;
 	CRect           trash_rect;
 	RPObjectPixmap* trash_bmp;
-	bool            is_trash_over( const rp::point& global_chart_pos ) const { return trash_rect.PtInRect( CPoint(global_chart_pos.x, global_chart_pos.y) ) ? true : false; }
+	bool            is_trash_over( const rp::point& global_chart_pos ) const { return trash_rect.PtInRect( CPoint(static_cast<int>(global_chart_pos.x), static_cast<int>(global_chart_pos.y)) ) ? true : false; }
 
 public:
 	RPObjectFlowChart( RPObject* parent );
